Use MalInt throughout rather than i64, and change it to isize
No particular reason for isize. I guess in principle it's better suited to the machine if you're not running a
64 bit arch (btw)?

+++ impls/rust-dmr/src/types.rs
@@ -10,7 +10,7 @@ use std::ops::{RangeFrom, RangeInclusive};
 pub type MalList = Vec<MalObject>;
 pub type MalVector = Vec<MalObject>;
 pub type MalMap = HashMap<HashKey, MalObject>;
-pub type MalInt = i64;
+pub type MalInt = isize;
 
 #[derive(Debug, PartialEq, Eq, Hash, Clone)]
 pub struct MalSymbol {
